requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
Django==3.1.3
django-appconf==1.0.4
django-ckeditor==6.0.0
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-simpleui==2021.3
django-taggit==1.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.2
Flask==2.0.1
idna==2.10
importlib-metadata==3.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonfield==3.1.0
lxml==4.6.3
Markdown==3.3.3
MarkupSafe==2.0.1
multidict==5.1.0
peewee==3.14.4
pilkit==2.0
Pillow==8.1.0
Pygments==2.7.4
PyJWT==2.1.0
pymongo==3.11.4
PyMySQL==1.0.2
pyquery==1.4.3
pytz==2021.1
redis==3.5.3
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.4.0
